+++ R/quantileResidualTests.R
@@ -19,7 +19,7 @@
 #' @inherit quantile_residuals references
 #' @seealso \code{\link{fitGMVAR}}, \code{\link{GMVAR}}, \code{\link{quantile_residuals}}, \code{\link{GIRF}},
 #'   \code{\link{diagnostic_plot}}, \code{\link{predict.gmvar}}, \code{\link{profile_logliks}},
-#'   \code{\link{LR_test}}, \code{\link{Wald_test}}
+#'   \code{\link{LR_test}}, \code{\link{Wald_test}}, \code{\link{cond_moment_plot}}
 #' @examples
 #' \donttest{
 #' ## These are long running examples that use parallel computing!